requirements.txt
paddlepaddle
paddleocr
opencv-python
Pillow
numpy
